third_party/WebKit/LayoutTests/fast/events/mouse-event-from-touch-source-device-event-sender.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<script src="../../resources/js-test.js"></script>
<script type="text/javascript">
description('Test that synthetic mouse events from touch events should have a sourceCapabilities and its firesTouchEvents should be true.');

function mouseHandler(event) {
    debug(event.type);
    shouldBeNonNull("event.sourceCapabilities");
    shouldBeTrue("event.sourceCapabilities.firesTouchEvents");
}

if (window.eventSender) {
    for (var evt of ['mousemove', 'mousedown', 'mouseup', 'click']) {
        document.addEventListener(evt, mouseHandler);
    }
  
    debug('Sending gestureTap which synthesizes mousemove, mousedown, mouseup and click');
    eventSender.gestureTap(10, 10);
    
} else {
   debug('This test requires eventSender.');
}

</script>
