third_party/blink/web_tests/fast/events/mouse-event-from-touch-source-device-event-sender-expected.txt

Test that synthetic mouse events from touch events should have a sourceCapabilities and its firesTouchEvents should be true.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Sending gestureTap which synthesizes mousemove, mousedown, mouseup and click
mousemove
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is true
mousedown
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is true
mouseup
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is true
click
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is true
PASS successfullyParsed is true

TEST COMPLETE
